Fail CI if any test fail
Before this change the command "make test" would return status code 0
even though a test failed and that was because the `popd` would always
succeed and that is the status code that would be the end result for
that target. Using `python -m` supposedly omits the need for having to
move into the test folder.