[api] Add proxy endpoint
This will serve as the general endpoint to redirect the user forward to
subpages of the application. It is necessary because we need to make
sure that a token exists in order to make any calls to either of the
APIs. So if the user does not have a valid token then this endpoint will
make sure that it receives one before doing the actual request.